+++ js/textbox/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gradio/textbox
 
+## 0.6.7
+
+### Dependency updates
+
+- @gradio/atoms@0.7.6
+- @gradio/utils@0.5.1
+- @gradio/statustracker@0.7.1
+- @gradio/icons@0.6.0
+
 ## 0.6.6
 
 ### Dependency updates